websocketsonly parameter to block non-websocket requests
If Nginx is used to load-balance websocket connections to multiple OMERO servers and someone makes a non-websocket request to the location OMERO will drop the connection, leading Nginx to think the server has failed and therefore remove it from the loadbalancing pool.

This adds a new option `websocketsonly` which if set will insert a check into the proxied location so non-websocket requests result in a HTTP status 400 (Bad Request client error).