Run restorecon in multithreaded mode
In the context of a production IDR deployment where the Nginx cache
is restored from a previous production environment, the restorecon
task is one of the biggest bottlenecks of the execution given the
number of files - in the order of millions.
This change sets the restorecon multithreaded option to 0 to create
as many threads as there are available CPU cores.